Moved tests from rethinkdb test_run_query_util to test_connection
